Review Rating/Likes done on both ends (untested)
Data for review ratings/likes (whatever you want to call it) is obtained
by piggybacking off of the cache data endpoint. No matter what, an extra
DB query needs to be made for this endpoint, so might as well return all
the liked/disliked data irregardless and let the front end handle the
logic.